Set the surname of the test advocate
One of the tests flickers because it is searching for the advocates surname
and expecting nothing to be returned but the surname is randomly chosen and
'Smith' is used as a defendant name. The solution is to fix the name of the
advocate to ensure that a result is not found.