Optimize query preparation
Journeys postgres DB function is dropped and replaced
on every call. This is an unneccessary overhead most of the
time as only changes to the function signature or body
would require a drop and replace respectively.

This slows tests and there are race conditions running
on production(-like) load balanced servers where this can
result in errors such as below. I believe this is due to
trying to replace a function in use. It is unclear what the
DROP command could do in such instances.

```
Error: PG::UniqueViolation: ERROR:  duplicate key value violates unique constraint "pg_proc_proname_args_nsp_index"
DETAIL:  Key (proname, proargtypes, pronamespace)=(journeys, 23, 16403) already
```

This change means it will only create the function if
1. There is no function of that name (irrespective of signature)
2. or if the app setting flag `Settings.replace_journeys_func?` is set to true.
   see below.

The `Settings.replace_journeys_func?`
-------------------------------------

Add a settings flag to enable recreation of the `journeys` function

Rather than replacing the the function everytime or using migrations
to apply changes (a viable alternative*) to "lazily" create the function
when it does not exist BUT only replace it if we configure the app or
instances to do so.

We default replacing the function to false as it already exists
and will not require replacing unless changes are made to its
signature or body.

 * migrations could be used to replace the function as an alternative.
 Example (untested):
 ```
 class ReplaceJourneysFunction < ActiveRecord::Migration[6.1]
  include Stats::ManagementInformation::JourneyQueryable
  def up
    send(:recreate_journeys_func)
  end

  def down
   raise ActiveRecord::IrreversibleMigration, 'function change cannot be undone!'
  end
end
```