Fix selenium deprecation warning
Fixes this warning, which can be seen on any @javascript
feature test and in circle

example:

When Running this:
```
cucumber features/authentication.feature
```

You get this
```
WARN Selenium [DEPRECATION] [:browser_options] :options as a parameter for driver initialization is deprecated. Use :capabilities with an Array of value capabilities/options if necessary instead.
```