Add docker image scanning to test suite workflow
To run security scanners on the docker image and dockerfile
to identify vulnerabilites.

Custom project name for snyk scan: Name after repo for clarity,
plus explicitly define the that snyk should monitor the repo as
well (which is the default behaviour but this is more declarative)

Schedule a scan rather than running as part of pipeline because:

1. another possible bottle neck on testing pipeline
2. eats up circleci credits
3. relies on PRs to detect possible upstream changes
   that could introduce vulnerabilites.

Base snyk project name on repo only: To reuse images being
monitored in snyk rather than creating a new one for each branch.

We are only interested in master branch based images
anyway.