Workaround for govuk_uk_date_field multiparameter assignment errors
Handle `ActiveRecord::MultiparameterAssignmentErrors` for date
fields when supplying args such as 32/01/2021 or A/01/2021 or
1/13/2021.

Applies this to `certification` controller, which already errors in
production, plus the external_users/claims_controller form so that
we can progress case_details form step migration.

This should not have side affects since it only "cleans" params
that have an 'attribute_name(2i)|attribute_name(3i)' parameter
in the params of the request.

It handles deeply nested multiparameter dates using a recursive loop
over the entire params.

To apply to another controller we just need to add a `before_action`
such as below. The prepend was necessary for the claims controller
as other `before_actions` were beating it to punch and thereby encountering
the error.
```
  prepend_before_action :clean_multiparameter_dates, only: [:create, :update]
```

NOTE: Rails implicit parsing as Time or Time.zone.local bumps forward
certain dates automatically. This is default behaviour for rails but is
not necessarily desirable.
e.g.

31-09-2021 --> 01-10-2021
29-01-2021 --> 01-03-2021

We could prevent this behaviour here as well?!