Allow API claim creation for inactive users
We are considering inactivating/sofftly-deleting
users that have not been "active" in the last 6 months
for security reasons.

Since deactivating changes the email address currently (does it have to?)
submissions from the API would no longer be able to find those emails.

This then finds the email by removing the "deleted" suffix.

Note that this would allow API claim drafting for inactive users as
both a claim `creator` and claim `external_user` (owner)