Add fixed fee and misc fee locales files for error model
The `:rate` validator is within the base fee validator, this is shared with both the fixed fee and misc fee model,
therefore changing the `:rate` validator error message from a string to a symbol, has now affected the error presentor and its expected message.

This resolves the faling spec where the expected error message was incorrect.