Allow arguments to override default parameters
`super(**query.merge(api_key: user.api_key))` will always use the default value
for `api_key` even if a value is passed in in the `query` hash. By replacing
with `super(api_key: user.api_key, **query)` it is possible to override the
default. This behaviour isn't required at the moment but it is more intuitive.