CFP-532 Provider admin: Display disabled user status
When a provider admin views the list of users on the Manage users page,
there is currently nothing to indicate that a user has been disabled;
they appear with the Edit and Delete links as with an enabled user. This
may be confusing for admins.

This commit updates the view to display a disabled user as 'Inactive'.