CFP-577 Generate GOVUK compatible error attributes
In order to work correctly with the GOVUK Formbuilder, we need to
correctly generate errors for 'has-one' attributes. This commit amends
the BaseSubModelValidator class to extend existing 'has-many'
functionality to also cover 'has-ones'.

Most of the changes are in the validate_association_for method. This
method now:
- finds the associated record
- checks if the record is nil/destroyed/marked for destruction or valid
- if any of these apply, returns as there are no errors to generate
- otherwise, sets @result as false and calls copy_errors_to_base_record

This will need to be configured by developers working on later tickets
by adding the name of the association to the guard clause to actually
be effective.