CFP-577 Hide non-GOVUK compatible error attributes
Errors generated via nested attributes are not compatible with the
GOVUK Formbuilder. We are now also generating those errors via the
BaseSubModelValidator class and so do not need to display the original
errors. This commit amends the ErrorMessage::Presenter class to hide
those errors.

This will need to be configured by developers working on later tickets
to actually be effective.