CFP-543 Ruby 3.1: Fix failing rspec-mock call
Following the update to ruby 3.1,
spec/requests/case_workers/admin/case_workers_spec.rb:56 fails due to a
bug in the rspec-mock gem.

The call `and_raise(NoMethodError)` raises a `NoMethodError` which
includes the text `self.terminal_implementation_action = Proc.new { raise(*args) }`
in the message. This is internal `rspec-mock` code which seems to be
bubbling up into the error by mistake, causing our test to fail. Other
exception types don't seem to be affected, so this commit switche to
mocking using `ArgumentError` which allows the test to pass.

The problem with the gem will be raised as an issue with the maintainers.