Remove TransferBrain#details_combo_valid?
This method does not appear to be used anywhere and it is not clear what its
purpose is.

The method calls TransferBrain/DataItemCollection.instance.detail_valid? so
the tests have been moved into spec/models/claim/transfer_brain/data_item_collection_spec.rb
instead. Some of the old tests were testing for transfer details combinations
that are not valid and do not in fact occur in production so I have removed
them completely.