CFP-624 SPIKE: Prevent ECNP on CLAIR LGFS Final Fees
Adds extra validation to `representation_order_validator` to check
whether a claim is for an Elected Case Not Proceeded, and if so
whether the representation order date falls after the start of the
CLAIR fee scheme.

If both these things are true, an error is displayed and the user
cannot continue.