CFP-627 Create script to add LGFS fee scheme 10 (CLAIR)
Creates a script that adds the new LGFS fee scheme 10 for
the CLAIR changes to be introduced on 30/09/2022.

* Adds a new seeder in `db/seeds/lgfs_scheme_10.rb` and adds it to
`db/seeds.rb`, so that databases are automatically seeded with the new
fee scheme;
* Creates a new rake task `lib/tasks/lgfs_scheme_ten.rake` which runs
code in a new seed schema `db/seeds/schemas/add_lgfs_fee_scheme_10.rb`
which controls the bulk of the process to create and rollback the fee scheme;
* Adds a start date for the fee scheme to `config/settings.yml`;
* Updates LGFS records in `lib/assets/data/fee_types.csv` to link
fee types to the appropriate LGFS fee schemes.