Allow the use of future dates for claim creation
There are scenarios where it would be useful to be able to create a claim
with a representation order dated in the future - for example when testing
a new fee scheme.

This commit adds control over the validation of future dates using an
environment variable. Setting the ALLOW_FUTURE_DATES variable for a given
environment will allow the creation of claims using future-dated rep orders.