Rake task to create AGFS Fee Scheme 13
Creates a script that adds the new AGFS fee scheme 13 for
the CLAIR changes to be introduced on 30/09/2022.

* Adds a new seeder in `db/seeds/scheme_13.rb` and adds it to `db/seeds.rb`, so
that databases are automatically seeded with the new fee scheme;
* Creates a new rake task `lib/tasks/agfs_scheme_thirteen.rake` which runs
code in a new seed schema `db/seeds/schemas/add_agfs_fee_scheme_13.rb`
which controls the bulk of the process to create and rollback the fee scheme;
* Adds a start date for the fee scheme to `config/settings.yml`;
* Updates AGFS records in `lib/assets/data/fee_types.csv` to link
fee types to the appropriate AGFS fee schemes.