Use string for date instead of erb for CLAIR date
`Date.new(year, month, day)` formats the date by default as `yyyy-mm-dd` and
this is how the `config` gem expects dates to be formatted, so that
`Settings.agfs_scheme_13_clair_release_date` is an instance of `Date`. The
default formatting of a date is changed in `config/initializers/date_time.rb`
into a format that the `config` does not understand and so
`Settings.agfs_scheme_13_clair_release_date` becomes a `String` instead. The
result of this is that in a development environment, where the config is
reloaded with each request, the type of these fields is non-deterministic.

There is no particular reason why the dates need to be generated using erb so
all the dates in the settings file could be changed into `yyyy-mm-dd` strings.
However, this should be done separately and not part of the CLAIR work.