CFP-630 Add final claim ECNP signpost
As a result of changes being introduced by CLAIR in AGFS fee scheme 13
and LGFS fee scheme 10, it will no longer be possible to claim ECNP fees
for  final claims.

To prevent users claiming these fees, a warning will be displayed on the
case type page after the start of the new schemes. This commit adds a
method to the ClaimsHelper class to control the display of that warning
based on date and claim type.