Fix ordering of fee scheme seeding
The ordering of the seeding of agfs fee scheme 13 and lgfs fee scheme 10
causes failures when seeding offence data in new databases due to
dependencies on the offence_id.

This commit fixes the ordering so that offences are seeded successfully.
We may want to change this functionality in future to prevent this
problem occurring again.

This commit also fixes a small typo in the agfs fee scheme 13 seeder.