SPIKE: Create change of monarch feature flag
The purpose of this commit is to demo the feature flag functionality,
not to implement any realistic monarch-related changes.

Uses the existing feature flag class lib/feature_flag.rb to create a
new flag and use it in code to control the display of some text.

Pros: This is simple and uses existing code.

Cons: It appears to be very simple and as it is controlled by config/settings.yml
doesn't easily allow different environments to be configured differently.

It also requires a release to update config/settings.yml in order to
remove the flag to enable a feature.