SPIKE: Create change of monarch feature flag
The purpose of this commit is to demo the feature flag functionality,
not to implement any realistic monarch-related changes.

Simply adds a new key to config/settings.yml which can be referenced in
code. The value assigned to this key can be read from an evironment
variable allowing for finer-grained control over environments.

Pros: This is simple and uses existing functionality. Can be configured
to enable/disable a feature differently on different environments.

Cons: As flags are configured in code, it requires a release to
to update config/settings.yml in order to enable/disable a feature.