Add new feature flag functionality
The purpose of this commit is to demo the feature flag functionality,
not to implement any realistic monarch-related changes.

Creates a new, more complex approach to feature flags. Adds a Feature
model which allows the flag to be persisted to the database. Only one
Feature object will ever exist, with a boolean column representing each
feature flag.

Pros: Allows flags to be set on an environment-by-environment basis. Does
not require a release to enable/disable a feature - they can be updated
by developers directly in the database, or potentially through the UI.

Cons: Requires a bit more work to initially implement, and to maintain.
Adding a new feature requires a database migration to add an extra column
to the database table.