Return body from #post_to_endpoint
For consistency with `get_dropdown_endpoint`, return the response body from
`post_to_endpoint` instead of the full response. This makes sense ans the
return value, whenever it is used, is always passed to `JSON.parse`, which
interprets it as a string and for a `RestClient::Response` instance this is
identical to the response body.