CFP-750 Use default docker version in CircleCI jobs
CircleCI is updating the `setup_remote_docker` feature. This may lead to
job failures where an old version of docker is specified for use with
the feature.

The version of docker specified in `.circleci/config.yml` (20.10.7) is on the list
of at-risk docker versions.

This commit removes the version flag, causing `setup_remote_docker` to
instead use the default docker version (20.10.17).