Remove 'pending' from timezone sensitive test
Reverts the change made in https://github.com/ministryofjustice/Claim-for-Crown-Court-Defence/pull/4773/files
to allow a timezone sensitive test to continue to pass during BST. Now
we are back in GMT the temporary change can be removed.

This will need to be re-added in March 2023 if we do not fix this
issue before.