Add Jasmine-jQuery argsToArray method
Jasmine-jQuery relies on an internal method in the Jasmine library
(jasmine.util.argsToArray). That method has now been removed from Jasmine:

https://github.com/jasmine/jasmine/commit/e2e2275d41f7a380a36b0855ada19ac5e7a9fa02

As a result failures occur in Jasmine-jQuery.

This commit adds a new function (argsToArray) to spec/javascripts/helpers/jasmine-jquery.js
which replicates exactly the functionality of jasmine.util.argsToArray.

This is far from ideal, as adding a new version of Jasmine-jQuery would
overwrite this change. However Jasmine-jQuery hasn't been updated in > 5
years so the risk is low (although the fact it is no longer maintained is
in itself is an issue).