Fix failing Jasmine tests
There are a number of Jasmine tests in the format

    xdescribe('...blockDisabledLinkClick', function () {
      // TODO
    })

The latest version of Jasmine no longer allows a describe() without a
child (an it() or describe()). This applies here even though the test
is skipped using xdescribe, resulting in failures.

This commit replaces the above format with

    xdescribe('...blockDisabledLinkClick', function () {
      xit('TODO')
    })

which allows the tests to be successfully skipped.