Temporarily fix REDIS_URL
The username needs to be removed from the redis url when upgrading the redis
gem from 4 to 5. The url is set in the REDIS_URL and this will be updated in a
separate PR for cloud-platform-environments, after which this temporary fix can
be reverted.

(WIP) Experiment before updating kubenetes secrets