Delete _nav_tabs partial
Snyk is reporting four XSS vulnerabilities in
app/views/case_workers/claims/_nav_tabs.html.haml. This file does not
appear to actually be in use; all references to it were removed from
code in 2015(!).

Because it's not in use, it's not actually vulnerable, but to keep Snyk
happy and tidy up the codebase, this commit deletes the file.