CTSKF-68 Disable feature flag in sandbox and staging
The CLAIR Contingency features are not yet ready to be visible to users
in the api-sandbox environment. This commit disables the
main_hearing_date feature flag in that environment to give us time to
communicate changes to software vendors before the changes are released
to them.

It also disables the feature flag in staging to keep consistency between
that environment and production.