Rubocop: RSpec/DescribeClass
RSpec/DescribeClass enforces the top level describe having the class or module
as the description. This is not relevant for view, request and API integration
specs.

View and request specs are listed in the default settings to be excluded but
these might be getting overridden by the AllCops setting at the top.
See https://docs.rubocop.org/rubocop-rspec/cops_rspec.html#configurable-attributes-6

Other files are listed individually to be fixed. For some reason these do not
when listed in .rubocop_todo.yml