Fix text alignment on claim summary screen
In the Defendants section of the summary screen for a submitted claim,
the text 'Judicial apportionment' wraps inelegantly, with the last 't'
appearing on a new line.

This commit improves the appearance by slightly increasing the width of `govuk-summary-list__key`
in the `app-card--defendant` element to allow the text to appear correctly.