Allow CircleCI to correctly split cucumber tests
CircleCI is configured to split cucumber tests by timings to allow for
more efficient parallelisation:

`circleci tests split --split-by=timings --timings-type=filename`

However this does not work correctly. All files result in the message
`No timing found for "<filename>"` when run.

This is occurring because the test results are output using the `json`
formatter, which is not supported by CircleCI for splitting by timings
(and additionally has now been deprecated).

This fixes the problem by:

* switching from using the `json` formatter to using the `junit` formatter,
which is required by
https://circleci.com/docs/use-the-circleci-cli-to-split-tests/#junit-xml-reports;
* adding the `fileattribute=true` flag to include filenames in the xml
output, which is needed for splitting accurately;
* writing the xml output to a directory rather than a file (a requirement
of the junit formatter).