CTSKF-128 Update ECNP Final claim validation
Claims that fall under CLAIR contigency rules (where the
representation order date is on or after 17 September 2020 with a main
hearing date on or after 31 October 2022) need to have the same
validation for Elected Cases Not Proceeded (ECNP) as claims that fall
under CLAIR rules (where the representation order date is on or after
30 September 2022).

This commit amends the logic in Claim::BaseClaimValidator to also include
the main hearing date in the validation. It also updates the wording of
the error displayed when the validation fails and uses feature flags to
switch betwen the different versions.