CTSKF-254 Fix filtering in FeeSchemeFactory::Base
The Base FeeSchemeFactory filters the correct fee scheme for a claim
based on an array of date ranges defined in the classes that inherit
from it (ie FeeSchemeFactory::AGFS and FeeSchemeFactory::LGFS).

The last item in each of those arrays is an open range, eg
`Settings.agfs_scheme_13_clair_release_date..`

Selection of the correct fee scheme is done using the `.include?`
method:

`return filter[:scheme] if filter[:range].cover?(@representation_order_date)`

When used on an open range, `include?` iterates infinitely and never
returns a result. See https://bugs.ruby-lang.org/issues/19418.

This occurs in CCCD on the very rare occasions that a representation
order date is used which predates the earliest date allowed for a fee
scheme (ie 01/04/2012). This results in the application hanging.

To fix this, add an end date of Time.zone.today to all currently
endless date ranges.