CTSKF-260 Add validation to main hearing date
The main hearing date field currently allows users to enter dates
in the format dd/mm/yy which results in dates such as 01/01/0023 being
saved to the database. This causes errors on injection to CCR and CCLF.

This commit adds validation consistent with
other date fields in the application to prevent dates older than 10
years being entered.