Add request spec for API swagger docs
The PR to remove webpacker from CCCD caused a failure in production
due to the interactive API documentation at `/api/documentation` not
rendering. We neglected to manually test the page and there is no
automated test that the page is working.

This commit adds a small request spec to test that the documentation
landing page is available.