Use strong params in management information controller
Snyk has a flagged an open redirect vulnerability in
CaseWorkers::Admin::ManagementInformationController due to the use of
unsanitized request parameters in a redirect_to.

That controller already implements strong parameters but they are not
always used in the code. This commit replaces direct access to the
parameters (eg `params[:report_type]`) with use of the strong parameters
(eg `report_params[:report_type]`).

(This is isn't strictly necessary as there is a before hook which
validates the :report_type parameter before it is used, which mitigates
the vulnerability, but this change is more idiomatic and keeps Snyk happy.