Log the inputs for API requests
Previously, the input data was only being logged when the fields were received
as form data, such as when requests are sent via the Swagger docs. Most(?)
vendor software, however, sends a JSON body and so this is not logged.

To log this, the `api.request.input` is logged at the end of the request (it is
set during the processing of the requests so it is blank at the start). To make
sure this appears in the log, the post-request log is now always written even
if the response is blank.