Trigger keyup event after change
If a misc fee with a fixed quantity exists for a claim then the quantity field
is set on load. To make sure that the net amount is calculated correctly there
needs to be an event to trigger the `feeQuantityChange` method in
FeeCalculator.UnitPrice.js.

A check is added before updating the quantity to the default and setting the
field as read only to avoid an infinite loop; changing the quantity causes the
rate to be fetched from Fee Calculator and then updating the rate causes the
quantity to be reset.