Remove unused files from docker image
The `builder` docker image used in the first stage of the multi-stage
docker build process contains unnecessary files which will be copied over
to the final image.

To reduce the size of the resulting image, this commit removes the
bundler cache and C source/object files, and removes the node_modules
directory from the application source.