Update Alpine dependencies before installation
At present we use apk to install Alpine dependencies early in the
build process and then call `apk --update-cache upgrade` at the end of
the process to check for newer versions of the dependencies.

This feels a bit back-to-front, so this commit moves the upgrade statement
immediately before the installation.

It also tidies up the list of dependencies a bit by alphabetising them.