CTSKF-318 Add Additional Prep Fee to fee_types.csv
This fee is introduced in AGFS Fee Scheme 15. The fee can be claimed on
final, hardship and supplementary claims for all trial, cracked trial,
retrial, and cracked before retrial cases with a representation order
date on or after 17/04/2023.

The Statutary Instrument laid before Parliament states that the fee is
additionally restricted to claims with a main hearing date on or after
16/05/2023, however this is a legislative/administrative statement only
due to an error in the previous SI, and is not being implemented in CCCD.
All claims in AGFS Fee Scheme 15 (ie with a representation order date on
or after 17th April 2023) are eligible for the fee.