CTSKF-318 Add API release note for Additional Prep Fee
Provides details of new Additional Preparataion Fee for software vendors.

Although the linked SI states that the fee is only claimable for cases
with a main hearing on or after 16th May 2023, this a legislative thing
only and is not being implemented in CCCD. All claims in fee scheme 15
(ie with a representation order date on or after 17th April 2023) are
eligible for the fee. The release note reflects the implementation, not
the SI.