Fix API smoke tests by defaulting to QC advocate role
The API smoke tests select the first available advocate role. Previously
the list of advocate categories was:

  - QC
  - Led junior
  - Leading junior
  - Junior alone

and so QC was always selected.

Commit c5658a8f47 changed the order of these categories to:

  - Junior alone
  - Leading junior
  - Led junior
  - QC

which means that Junior alone is now always used as the advocate category.

Junior alone was removed as an advocate category after fee scheme 9, so
when the API smoke tests test a fee scheme 10 claim, the test fails due
to the use of an invalid advocate category.

This commit forces AdvocateClaimTest::Base to use QC as an advocate
category, essentially returning to the previous format, so that tests
now always pass.