Add env variable and update labels for the worker pods in k8s
Distinguish the app and worker pods by updating the worker pods label to cccd-worker.
This is so that we can find a way to get the service to correctly route requests without failing with 5xx errors intermittently.
The service sometimes routes API requests to the sidekiq worker pods instead of the application pods, which causes a 5xx error (usually 502)
Now that the worker pod doesn't have a label matched with the service, the service will no longer send http requests to that pod but instead to the cccd-app pods

Add new env variable for Caseworker API calls:
Before, the caseworker api calls were made via the GRAPE_SWAGGER_ROOT_URL but it makes more sense
to decouple it: one for the swagger docs and another for the api calls

The new environment variable is called CASEWORKER_API_URL

Set the value of CASEWORKER_API_URL to the k8s service url cccd-app-service . This is an internally routed url.

Not only does this decouple the Caseworker API call from the Swagger Documentation, it also fixed a CORs issue. When we tried to use the same url but updated it with
cccd-app-service, it would fail on the swagger documentation because of CORs. The RestClient call to cccd-app-servivce had a different domain name to our usual service url.
This seperation of environment variables prevents the need to work around the CORs issue here.

Update all environments with the above.

Update usage of GRAPE_SWAGGER_ROOT_URL where the caseworker api is being called to use CASEWORKER_API_URL.