Retry entering autocomplete options
In some cases the autocomplete finds 'No results found'. This appears to be a
timing issue that will only be seen in the testing environment, where the list
of options in the autocomplete list is being checked immediately after the
input has been entered. This can happen before the autocomplete module has
updated, which is done every 100 milliseconds.
but to make the tests more reliable this can by tried up to three times.