Add margin below caseworker summary expenses table
All govuk tables should have a margin-bottom value of 30px (equivalent
to 6 govuk spacing units). The use of the datatables library on the
expenses table on the caseworker claim summary screen overrides this and
sets the margin to 0. This means there is no spacing between the table
and the next element on the screen.

This change adds extra styling to app/webpack/stylesheets/components/_table.scss
to explicitly add the necessary margin to the expenses table.