Fix alignment of numeric cells in caseworker expenses table
GDS style guidelines state the numeric values in tables (and the headers
of those columns) should be right-justified. The use of the
datatables library in the expenses table on the caseworker claim summary
screen is causing this styling to be lost.

This adds extra styling to app/webpack/stylesheets/components/_table.scss
to explicitly set `text-align` to `right` on that table.