Rubocop: Style/ReturnNilInPredicateMethodDefinition
The default auto-correct is to return `false` and this makes sense for methods
returning a true/false value and this, by convention, is the case for methods
ending in `?`. `error_class?`, however, returns a class name that is used in
html elements and so `nil` (or possibly `''`) makes more sense.