CTSKF-436 Remove ECR deletion jobs
The move to using short-lived credentials for accessing ECR means that
the existing `delete-ecr-images` GitHub Action and `clean_ecr` cronjob
will no longer work. This is because the AWS role assumed for IAM
purposes follows the principal of least privilege and does not allow
deletion of images.

https://mojdt.slack.com/archives/C57UPMZLY/p1686037156333669?thread_ts=1686036923.591069&cid=C57UPMZLY

Instead, lifecycle policies will be added to the Cloud Platform config
to control the deletion of old images. This commit removes the two jobs
from the CCCD repo.