CTSKF-436 Improve tagging of prod docker images
When a docker image is built from the master branch, for deployment to
production, it is tagged twice, once as `app-<SHA>` (as all images
are) and once as `app-latest`, and pushed to ECR. The next time a
production deploy occurs the same thing happens. This results in the
first image being untagged as `app-latest`, replaced by the second image.

The first image still exists in ECR, but is very difficult to identify
as a production image. If we needed to rollback to a previous image we
would need to identify it via the SHA from CircleCI logs. This lack of
meaningful tagging also makes it difficult to define an ECR lifecycle
policy to retain production images.

This change amends the build process to add an additional tag to each
image built from the master branch. This maintains the current `app-latest`
format, to signify a production image, but appends extra data to uniquely
identify the image. This means that when a new `app-latest` is built,
previous `app-latest` images can still be identified.