Upgrade to yarn v3.6.1
We are starting to experience issues updating javascript dependencies.
These issues can be resolved by moving to the latest version of yarn.

The upgrade is done following instructions here:

https://yarnpkg.com/getting-started/migration

It checks in the new .yarn/cache folder, which adds a *lot* of files to
the repo, but has follows the recommendation here:

https://yarnpkg.com/getting-started/qa#which-files-should-be-gitignored

and has benefits detailed here:

https://medium.com/inato/make-your-ci-faster-and-improve-developer-experience-upgrade-to-yarn-2-with-or-without-939524d9bef2

Note: this does not enable the new Plug'n'Play strategy as doing so
causes errors when starting the server. This can be addressed in a later
change.