CTSKF-373 Fix failing state transition validations
The automated overnight stale claim archiver fails each night with 80
claims that cannot be archived. These 80 claims all fail due to
validation errors.

Claims should not be validated when transitioning state. Code in
the StateMachine model disables validation, but needs to be called by
the claim models being validated. This is done for all advodate claims
and the LitgatorClaim model, but not for litigator interim, hardship or
transfer claims.

All 80 claims currently affected by this issue are litigator interim
claims.

This commit updates the validation rules on the litigator InterimClaim,
LitigatorHardshipClaim and TransferClaim models, bringing them inline
with the LitigatorClaim model.