Remove unused `deleted` state
The StateMachine module defines a `deleted` state, which is not used
anywhere in the system. This may have been intended as a 'soft delete'
state but this is now handled by setting the `deleted_at` attribute.

When claims are finally deleted, they are removed from the database
using the rails `destroy` method, so this state is redundant and is
removed here.