Fix inaccurate state_machine tests
Two tests in `spec/models/claims/state_machine_spec.rb` claim to test the
allocation of claims that are in a state of `redetermination` or
`awaiting_written_reasons` however the test setup means that they are
`submitted` when the tests run.

This commit puts the claims in the correct states before the tests run.