Remove check on offence fee scheme for agfs reform
As noted in the TODO comment, the offence check should be removed as it is
already handled by `FeeScheme.for_claim`, as the correct fee scheme is found
based on the offence when the earliest representation order date is not set.
This situation can occur when a claim is submitted via the API as the offence
can be set before the defendants.

It is now necessary to remove this check as the LGFS fee scheme 10 offences are
identical to the AGFS and LGFS fee scheme 9 offences in the Offence model. This
means that a check that a particular offence is in scheme 10 may erroniously
be true for an AGFS fee scheme 9 offence, as it is also the offence for LGFS
fee scheme 10.