Task to merge AGFS fee scheme offences
Three new tasks are created:

db:agfs_scheme_offences_clean:status: Display the status of AGFS offences
linked to each fee scheme. For each fee scheme the total number of offences,
the first and last ids and the number of claims with this offence are
displayed. Additionally, a file for each fee scheme is created listing all the
claim ids for each offence grouped by whether they were created via the web,
the api or the api and subsequently edited on the web. The purpose of this file
is to be able to compare after a rollback.

db:agfs_scheme_offences_clean:merge: Remove all the offences linked to fee
schemes 12 and 13. The fee scheme 11 offences are linked to fee schemes 12 and
13. All the claims with offences linked to fee schemes 12 and 13 are updated to
link to the corresponding offence linked to fee scheme 11.

db:agfs_scheme_offences_clean:rollback: Recreate the offences for fee schemes
12 and 13 and unlink these fee schemes from the fee scheme 11 offences. All the
claims are checked for their fee scheme based on representation order dates and
their linked offences are updated as appropriate.