Fix Rails/I18nLocaleTexts offence in SessionsController
SessionsController contains hard-coded text that is displayed when a
user logs out of CCCD. This should be defined in a translation file
instead.

As sessions are handled by devise, it can be added to `devise.en.yml`
which means (after a slight change to the structure of that file) it is
automatically picked up and rendered by devise on logout, so there is
no longer need to explicitly render the text in the controller.