Fix Rails/I18nLocaleTexts offence in Message model
Moves error text from the model to `config/locales/en.yml`.

By doing this, and ensuring that the text is correctly defined in the
translation file with a key in the format
`en.activerecord.errors.models.<model_name>.attributes.<attribute_name>`,
we no longer need to explicitly reference the error in the model - the
magic of rails ensures that the correct message is used.