Fix Rubocop RSpec/RepeatedExampleGroupBody offences
Three tests are being flagged by rubocop because they are set up
incorrectly and are effectively duplicates of other tests. This commit
fixes them so that rubocop no longer complains and, more importantly,
the tests are actually accurate.

* Fixes the test for `POST #create` in `spec/controllers/external_users/admin/external_users_controller_spec.rb`
so that it actually tests `POST #create`, not `GET #new`.

* Fixes `spec/presenters/fee/fixed_fee_presenter_spec.rb` so that an
advocate claim is used for the AGFS tests and a litigator claim is
used for the LGFS tests.

* Fixes `spec/services/claims/fee_calculator/graduated_price_spec.rb` so
that tests for `Number of prosecution witnesses (NPW) fee` use the NPW
fee (`BANPW`) and not the PPE fee (`BAPPE`).